Ediciones | Carga de productos en pagina principal
- Acciones de edicion y falta terminadas, falta validación de lista de insumos.
- Muestra productos en la pagina inicial.

+++ BomboProyect/Models/Insumos.cs
@@ -44,7 +44,7 @@ namespace BomboProyect.Models
         //public List<DetProducto> DetProductos { get; set; }
 
         [NotMapped]
-        [Range(0.0, Double.MaxValue, ErrorMessage = "El campo debe ser mayor a 0")]
+        [Range(0, Int32.MaxValue, ErrorMessage = "El campo debe ser mayor a 0")]
         public double CantProduc { get; set; }
     }
 }
